Resposta modelo architecture initialized
Made a view that shows a relation of teachers and courses. It only shows
one class of a teacher if he teaches more than one class (usually this
happens so it's a good thing to do).
The filtering view is just a skel and needs to be implemented

Hint: When showing the answers remember to also group all the
'atribuição' of the same teacher to the same sheet of answers.